Workaround: ignore incorrect variable handling for robotframework
This is a temporary workaround for parsing Robot Framework results.
Variable handling is currently ignored, as there is no clean solution yet
to fully support variables (VAR) in different scopes.

In RF 7.x, a variable (VAR) can exist in the scope of Suite, Keyword, or Test.
The current code only handles keyword variables, which can lead to a
RuntimeError due to mismatched stack types.

Signed-off-by: Oliver Sedlbauer <osedlbauer@tdt.de>